+++ src/main/java/com/iem/FilmRentalStore/repository/ActorRepository.java
@@ -4,6 +4,6 @@ import com.iem.FilmRentalStore.entity.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+public interface ActorRepository extends JpaRepository<Actor, Short> {
     List<Actor> findByFirstName(String firstName);
 }
