+++ src/main/java/com/iem/FilmRentalStore/repository/FilmTextRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilmTextRepository extends JpaRepository<FilmText, Integer>, JpaSpecificationExecutor<FilmText> {
+public interface FilmTextRepository extends JpaRepository<FilmText, Short>, JpaSpecificationExecutor<FilmText> {
 }
